+++ crypto_sign/gemss-red-128/avx2/mixEquationsMQS_gf2.c
@@ -49,6 +49,6 @@ void PQCLEAN_GEMSSRED128_AVX2_mixEquationsMQS8_gf2_right(mqsnv8_gf2m pk, cst_mqs
     }
 
     for (i = 0; i < (5); ++i) {
-        pk[i] = pk_last[NB_WORD_GF2m - 1] >> (i << 3);
+        pk[i] = (uint8_t)(pk_last[NB_WORD_GF2m - 1] >> (i << 3));
     }
 }
